chore: refactor menuDataGenerator -> createMenuData
